010_sql_database_agent_app/requirements2.txt
streamlit
sqlalchemy
pandas
openai
langchain
langchain-openai
langchain-community
google-generative-ai
git+https://github.com/business-science/ai-data-science-team.git --upgrade
requests
